Engine: remove inv item limit, store item quantities in a map
NOTE: the technical limit is still 64k individual items because the "number of items" field in game data is int16

+++ Common/ac/gamesetupstructbase.h
@@ -50,7 +50,7 @@ struct GameSetupStructBase
     int               numviews = 0;
     int               numcharacters = 0;
     int               playercharacter = -1;
-    int               numinvitems = 0;
+    uint16_t          numinvitems = 0u;
     int               numdialog = 0;
     int               numdlgmessage = 0;    // [DEPRECATED]
     int               numfonts = 0;
